requirements.txt
tkinter==0.1.0
python-chess==0.31.4
# Stockfish (Install separately)